+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/cloud/AppEngineDeploymentRunConfigurationEditor.form
@@ -191,7 +191,7 @@
           <text value="Project:"/>
         </properties>
       </component>
-      <component id="84898" class="com.google.cloud.tools.intellij.elysium.ProjectSelector" binding="projectSelector">
+      <component id="84898" class="com.google.cloud.tools.intellij.resources.ProjectSelector" binding="projectSelector">
         <constraints>
           <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
