+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/debugger/ui/CloudAttachDialog.form
@@ -28,7 +28,7 @@
           <text resource-bundle="messages/CloudToolsBundle" key="clouddebug.target"/>
         </properties>
       </component>
-      <component id="30c54" class="com.google.cloud.tools.intellij.elysium.ProjectSelector" binding="elysiumProjectSelector">
+      <component id="30c54" class="com.google.cloud.tools.intellij.resources.ProjectSelector" binding="elysiumProjectSelector">
         <constraints>
           <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
