+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/debugger/ui/CloudDebugRunConfigurationPanel.form
@@ -10,7 +10,7 @@
     </properties>
     <border type="none"/>
     <children>
-      <component id="30c54" class="com.google.cloud.tools.intellij.elysium.ProjectSelector" binding="elysiumProjectId">
+      <component id="30c54" class="com.google.cloud.tools.intellij.resources.ProjectSelector" binding="elysiumProjectId">
         <constraints>
           <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
