+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/debugger/ui/ProjectDebuggeeBinding.java
@@ -23,8 +23,8 @@ import com.google.api.services.clouddebugger.v2.model.ListDebuggeesResponse;
 import com.google.cloud.tools.intellij.CloudToolsPluginInfoService;
 import com.google.cloud.tools.intellij.debugger.CloudDebugProcessState;
 import com.google.cloud.tools.intellij.debugger.CloudDebuggerClient;
-import com.google.cloud.tools.intellij.elysium.ProjectSelector;
 import com.google.cloud.tools.intellij.login.CredentialedUser;
+import com.google.cloud.tools.intellij.resources.ProjectSelector;
 import com.google.cloud.tools.intellij.util.GctBundle;
 import com.google.common.base.Strings;
