+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/resources/BaseGoogleLoginUi.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.elysium;
+package com.google.cloud.tools.intellij.resources;
 
 import com.google.cloud.tools.intellij.login.ui.GoogleLoginIcons;
