+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/resources/ProjectSelectorItem.java
@@ -14,29 +14,48 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.elysium;
+package com.google.cloud.tools.intellij.resources;
 
+import com.intellij.ui.UI;
 import com.intellij.ui.components.JBLabel;
 
 import org.jetbrains.annotations.NotNull;
 
 import java.awt.Color;
-import java.awt.Font;
 
 import javax.swing.BorderFactory;
 import javax.swing.SwingConstants;
 
 /**
- * UI for the node that displays error information if an elysium call fails.
+ * Represents a single GCP project ui.
  */
-class ProjectSelectorErrorItem extends JBLabel {
+class ProjectSelectorItem extends JBLabel {
 
-  public ProjectSelectorErrorItem(@NotNull Color errorForeground) {
+  private Color textSelectionColor;
+  private Color textNonSelectionColor;
+  private Color hoverColor;
+
+  public ProjectSelectorItem(@NotNull Color backgroundNonSelectionColor,
+      @NotNull Color textSelectionColor, @NotNull Color textNonSelectionColor) {
     setBorder(BorderFactory.createEmptyBorder(2, 15, 2, 0));
     setOpaque(false);
     setHorizontalAlignment(SwingConstants.LEFT);
     setVerticalAlignment(SwingConstants.CENTER);
-    setFont(new Font(getFont().getFontName(), Font.BOLD, getFont().getSize()));
-    setForeground(errorForeground);
+    this.textSelectionColor = textSelectionColor;
+    this.textNonSelectionColor = textNonSelectionColor;
+
+    hoverColor = UI.getColor("link.foreground");
+    setBackground(backgroundNonSelectionColor);
+  }
+
+  public void initialize(String projectName, String projectId, boolean selected, boolean hovered) {
+    setText(projectName + " (" + projectId + ")");
+    if (selected) {
+      setForeground(textSelectionColor);
+    } else if (hovered) {
+      setForeground(hoverColor);
+    } else {
+      setForeground(textNonSelectionColor);
+    }
   }
 }
