+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/vcs/ChooseProject.form
@@ -26,7 +26,7 @@
           <grid row="1" column="1" row-span="1" col-span="3" vsize-policy="1" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
       </hspacer>
-      <component id="857e7" class="com.google.cloud.tools.intellij.elysium.ProjectSelector" binding="projectSelector" custom-create="true">
+      <component id="857e7" class="com.google.cloud.tools.intellij.resources.ProjectSelector" binding="projectSelector" custom-create="true">
         <constraints>
           <grid row="0" column="1" row-span="1" col-span="3" vsize-policy="2" hsize-policy="2" anchor="0" fill="3" indent="0" use-parent-layout="false">
             <minimum-size width="200" height="-1"/>
