+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/vcs/ChooseProjectDialog.java
@@ -16,8 +16,8 @@
 
 package com.google.cloud.tools.intellij.vcs;
 
-import com.google.cloud.tools.intellij.elysium.ProjectSelector;
 import com.google.cloud.tools.intellij.login.CredentialedUser;
+import com.google.cloud.tools.intellij.resources.ProjectSelector;
 
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.DialogWrapper;
