+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/vcs/CloneGcpDialog.form
@@ -57,7 +57,7 @@
         </constraints>
         <properties/>
       </component>
-      <component id="857e7" class="com.google.cloud.tools.intellij.elysium.ProjectSelector" binding="repositoryUrl" custom-create="true">
+      <component id="857e7" class="com.google.cloud.tools.intellij.resources.ProjectSelector" binding="repositoryUrl" custom-create="true">
         <constraints>
           <grid row="0" column="1" row-span="1" col-span="3" vsize-policy="2" hsize-policy="2" anchor="0" fill="3" indent="0" use-parent-layout="false">
             <minimum-size width="200" height="-1"/>
