+++ google-cloud-tools-plugin/testSrc/com/google/cloud/tools/intellij/resources/GoogleUserModelItemTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.elysium;
+package com.google.cloud.tools.intellij.resources;
 
 import com.google.cloud.tools.intellij.CloudToolsPluginInfoService;
 import com.google.cloud.tools.intellij.testing.BasePluginTestCase;
